Prefetch Fix
Issue: RabbitMQ appears to work fine, but it looks like it just cycles through the workers instead of properly distributing the load. To demonstrate this, I can take a long running process, then in another window run =(^)quack. If I just have two workers, it processes one request while the other is going. If I have 6 workers, it processes 5. After that it freezes, so it appears to simply be cycling.

Solution: Set prefetch count on server side to 1.